fix: tx dots for bad time ranges
when the price range has bad data (new installs have difference in ranges. this is
because historically only daily data is loaded into the db. however it then
collects data every 5 minutes so hourly data could be different for range of 3m for example.
the later dates could have hourly but the first dates dont).